Create README.md
An aspiring data analyst with a strong foundation in data visualization and Python. Currently pursuing degree in Data Science and data analytics with coursework covering topics such as predictive analysis and knowledge in programming languages such as SQL and R. Eager to apply my skills for the real-world applications and contribute to data decisions.